requirements.txt
####### requirements.txt #######
# jira api python client
jira
# PyJWT - lib for generating JWT tokens
PyJWT
# requests - lib for http calls
requests>=2.9.0
# pyyaml - lib for reading .yml files
pyyaml
# bitbucket api python client
stashy
